Yet another JVM (minor) fix (bootstrap)

+++ runtime/Jlib/input_port.java
@@ -7,7 +7,7 @@ public abstract class input_port extends obj
   public final String name;
   public int filepos = 0;
   public int pseudoeof = -1;
-  public int length = -1;
+  public long length = -1;
   public int bufsiz;
   public boolean eof = false;
   public boolean other_eof = false;
@@ -22,7 +22,7 @@ public abstract class input_port extends obj
   public input_port( final String name, final byte[] buf )
   {
     this.name= name;
-    this.bufsiz= buf.length;
+    this.bufsiz= (int)buf.length;
     buffer= buf;
     buf[0] = (byte)'\0';
   }
